+++ starlette/concurrency.py
@@ -1,19 +1,27 @@
 import functools
+import sys
 import typing
-from typing import Any, AsyncGenerator, Iterator
+import warnings
 
 import anyio
 
-try:
-    import contextvars  # Python 3.7+ only or via contextvars backport.
-except ImportError:  # pragma: no cover
-    contextvars = None  # type: ignore
+if sys.version_info >= (3, 10):  # pragma: no cover
+    from typing import ParamSpec
+else:  # pragma: no cover
+    from typing_extensions import ParamSpec
 
 
 T = typing.TypeVar("T")
+P = ParamSpec("P")
 
 
 async def run_until_first_complete(*args: typing.Tuple[typing.Callable, dict]) -> None:
+    warnings.warn(
+        "run_until_first_complete is deprecated "
+        "and will be removed in a future version.",
+        DeprecationWarning,
+    )
+
     async with anyio.create_task_group() as task_group:
 
         async def run(func: typing.Callable[[], typing.Coroutine]) -> None:
@@ -25,15 +33,9 @@ async def run_until_first_complete(*args: typing.Tuple[typing.Callable, dict]) -
 
 
 async def run_in_threadpool(
-    func: typing.Callable[..., T], *args: typing.Any, **kwargs: typing.Any
+    func: typing.Callable[P, T], *args: P.args, **kwargs: P.kwargs
 ) -> T:
-    if contextvars is not None:  # pragma: no cover
-        # Ensure we run in the same context
-        child = functools.partial(func, *args, **kwargs)
-        context = contextvars.copy_context()
-        func = context.run
-        args = (child,)
-    elif kwargs:  # pragma: no cover
+    if kwargs:  # pragma: no cover
         # run_sync doesn't accept 'kwargs', so bind them in here
         func = functools.partial(func, **kwargs)
     return await anyio.to_thread.run_sync(func, *args)
@@ -43,7 +45,7 @@ class _StopIteration(Exception):
     pass
 
 
-def _next(iterator: Iterator) -> Any:
+def _next(iterator: typing.Iterator[T]) -> T:
     # We can't raise `StopIteration` from within the threadpool iterator
     # and catch it outside that context, so we coerce them into a different
     # exception type.
@@ -53,7 +55,9 @@ def _next(iterator: Iterator) -> Any:
         raise _StopIteration
 
 
-async def iterate_in_threadpool(iterator: Iterator) -> AsyncGenerator:
+async def iterate_in_threadpool(
+    iterator: typing.Iterator[T],
+) -> typing.AsyncIterator[T]:
     while True:
         try:
             yield await anyio.to_thread.run_sync(_next, iterator)