+++ tests/middleware/test_base.py
@@ -3,8 +3,8 @@ import pytest
 from starlette.applications import Starlette
 from starlette.middleware import Middleware
 from starlette.middleware.base import BaseHTTPMiddleware
-from starlette.responses import PlainTextResponse
-from starlette.routing import Route
+from starlette.responses import PlainTextResponse, StreamingResponse
+from starlette.routing import Mount, Route, WebSocketRoute
 
 
 class CustomMiddleware(BaseHTTPMiddleware):
@@ -14,21 +14,23 @@ class CustomMiddleware(BaseHTTPMiddleware):
         return response
 
 
-app = Starlette()
-app.add_middleware(CustomMiddleware)
-
-
-@app.route("/")
 def homepage(request):
     return PlainTextResponse("Homepage")
 
 
-@app.route("/exc")
 def exc(request):
-    raise Exception()
+    raise Exception("Exc")
 
 
-@app.route("/no-response")
+def exc_stream(request):
+    return StreamingResponse(_generate_faulty_stream())
+
+
+def _generate_faulty_stream():
+    yield b"Ok"
+    raise Exception("Faulty Stream")
+
+
 class NoResponse:
     def __init__(self, scope, receive, send):
         pass
@@ -40,11 +42,22 @@ class NoResponse:
         pass
 
 
-@app.websocket_route("/ws")
 async def websocket_endpoint(session):
     await session.accept()
     await session.send_text("Hello, world!")
     await session.close()
+
+
+app = Starlette(
+    routes=[
+        Route("/", endpoint=homepage),
+        Route("/exc", endpoint=exc),
+        Route("/exc-stream", endpoint=exc_stream),
+        Route("/no-response", endpoint=NoResponse),
+        WebSocketRoute("/ws", endpoint=websocket_endpoint),
+    ],
+    middleware=[Middleware(CustomMiddleware)],
+)
 
 
 def test_custom_middleware(test_client_factory):
@@ -52,8 +65,13 @@ def test_custom_middleware(test_client_factory):
     response = client.get("/")
     assert response.headers["Custom-Header"] == "Example"
 
-    with pytest.raises(Exception):
+    with pytest.raises(Exception) as ctx:
         response = client.get("/exc")
+    assert str(ctx.value) == "Exc"
+
+    with pytest.raises(Exception) as ctx:
+        response = client.get("/exc-stream")
+    assert str(ctx.value) == "Faulty Stream"
 
     with pytest.raises(RuntimeError):
         response = client.get("/no-response")
@@ -61,30 +79,6 @@ def test_custom_middleware(test_client_factory):
     with client.websocket_connect("/ws") as session:
         text = session.receive_text()
         assert text == "Hello, world!"
-
-
-def test_middleware_decorator(test_client_factory):
-    app = Starlette()
-
-    @app.route("/homepage")
-    def homepage(request):
-        return PlainTextResponse("Homepage")
-
-    @app.middleware("http")
-    async def plaintext(request, call_next):
-        if request.url.path == "/":
-            return PlainTextResponse("OK")
-        response = await call_next(request)
-        response.headers["Custom"] = "Example"
-        return response
-
-    client = test_client_factory(app)
-    response = client.get("/")
-    assert response.text == "OK"
-
-    response = client.get("/homepage")
-    assert response.text == "Homepage"
-    assert response.headers["Custom"] == "Example"
 
 
 def test_state_data_across_multiple_middlewares(test_client_factory):
@@ -110,14 +104,17 @@ def test_state_data_across_multiple_middlewares(test_client_factory):
             response.headers["X-State-Bar"] = request.state.bar
             return response
 
-    app = Starlette()
-    app.add_middleware(aMiddleware)
-    app.add_middleware(bMiddleware)
-    app.add_middleware(cMiddleware)
-
-    @app.route("/")
     def homepage(request):
         return PlainTextResponse("OK")
+
+    app = Starlette(
+        routes=[Route("/", homepage)],
+        middleware=[
+            Middleware(aMiddleware),
+            Middleware(bMiddleware),
+            Middleware(cMiddleware),
+        ],
+    )
 
     client = test_client_factory(app)
     response = client.get("/")
@@ -151,9 +148,18 @@ def test_fully_evaluated_response(test_client_factory):
             await call_next(request)
             return PlainTextResponse("Custom")
 
-    app = Starlette()
-    app.add_middleware(CustomMiddleware)
+    app = Starlette(middleware=[Middleware(CustomMiddleware)])
 
     client = test_client_factory(app)
     response = client.get("/does_not_exist")
     assert response.text == "Custom"
+
+
+def test_exception_on_mounted_apps(test_client_factory):
+    sub_app = Starlette(routes=[Route("/", exc)])
+    app = Starlette(routes=[Mount("/sub", app=sub_app)])
+
+    client = test_client_factory(app)
+    with pytest.raises(Exception) as ctx:
+        client.get("/sub/")
+    assert str(ctx.value) == "Exc"